Change removal of extraction file to detect previously removed files

+++ src/sim/Abstract_Runnable_ClusterModel_Transmission.java
@@ -226,22 +226,22 @@ public abstract class Abstract_Runnable_ClusterModel_Transmission extends Abstra
 			pat_format = pat_format.replaceFirst("%d", Long.toString(sIM_SEED));
 			pat_format = pat_format.replaceFirst("%d", "(\\\\d+)");
 			Pattern pattern_sel_zip = Pattern.compile(pat_format + ".7z");
-			
+
 			File[] candidate = baseDir.listFiles(new FileFilter() {
 				@Override
 				public boolean accept(File pathname) {
 					return pattern_sel_zip.matcher(pathname.getName()).matches();
 				}
-			});						
-			// Replace with the unzip version if exist						
-			for(int f = 0; f < candidate.length; f++) {							
-				String rawFileName = candidate[f].getName().substring(0, candidate[f].getName().length()-3);				
-				File rawFile = new File(baseDir, rawFileName);				
-				if(rawFile.exists()) {
+			});
+			// Replace with the unzip version if exist
+			for (int f = 0; f < candidate.length; f++) {
+				String rawFileName = candidate[f].getName().substring(0, candidate[f].getName().length() - 3);
+				File rawFile = new File(baseDir, rawFileName);
+				if (rawFile.exists()) {
 					candidate[f] = rawFile;
-				}										
-			}						
-			
+				}
+			}
+
 			Pattern pattern_sel = Pattern.compile(pat_format + ".*");
 			ArrayList<Integer> timePt_Collection_fitlered = new ArrayList<>();
 			for (File f : candidate) {
@@ -325,10 +325,12 @@ public abstract class Abstract_Runnable_ClusterModel_Transmission extends Abstra
 
 		// Remove extracted file
 		for (File f : extracted_file) {
-			try {
-				Files.delete(f.toPath());
-			} catch (IOException e) {
-				e.printStackTrace(System.err);
+			if (f.exists()) {
+				try {
+					Files.delete(f.toPath());
+				} catch (IOException e) {
+					e.printStackTrace(System.err);
+				}
 			}
 		}
